remove beta pages
