+++ nuxt.config.js
@@ -58,7 +58,6 @@ export default {
     '~/plugins/sepaApi',
     '~/plugins/achApi',
     '~/plugins/mocksApi',
-    '~/plugins/beta/payoutsApi',
     '~/plugins/businessAccount/addressesApi',
     '~/plugins/businessAccount/balancesApi',
     '~/plugins/businessAccount/bankAccountsApi',
