--- pages/debug/beta/payouts/details.vue
@@ -1,80 +1,0 @@
-<template>
-  <v-layout>
-    <v-row>
-      <v-col cols="12" md="4">
-        <v-form>
-          <v-text-field v-model="formData.payoutId" label="Payout Id" />
-          <v-btn
-            depressed
-            class="mb-7"
-            color="primary"
-            :loading="loading"
-            @click.prevent="makeApiCall()"
-          >
-            Make api call
-          </v-btn>
-        </v-form>
-      </v-col>
-      <v-col cols="12" md="8">
-        <RequestInfo
-          :url="requestUrl"
-          :payload="payload"
-          :response="response"
-        />
-      </v-col>
-    </v-row>
-    <ErrorSheet
-      :error="error"
-      :show-error="showError"
-      @onChange="onErrorSheetClosed"
-    />
-  </v-layout>
-</template>
-
-<script lang="ts">
-import { Component, Vue } from 'nuxt-property-decorator'
-import { mapGetters } from 'vuex'
-import RequestInfo from '~/components/RequestInfo.vue'
-import ErrorSheet from '~/components/ErrorSheet.vue'
-@Component({
-  components: {
-    RequestInfo,
-    ErrorSheet,
-  },
-  computed: {
-    ...mapGetters({
-      payload: 'getRequestPayload',
-      response: 'getRequestResponse',
-      requestUrl: 'getRequestUrl',
-    }),
-  },
-})
-export default class FetchPayoutDetailsClass extends Vue {
-  // data
-  formData = {
-    payoutId: '',
-  }
-
-  required = [(v: string) => !!v || 'Field is required']
-  error = {}
-  loading = false
-  showError = false
-  // methods
-  onErrorSheetClosed() {
-    this.error = {}
-    this.showError = false
-  }
-
-  async makeApiCall() {
-    this.loading = true
-    try {
-      await this.$payoutsApiBeta.getPayoutById(this.formData.payoutId)
-    } catch (error) {
-      this.error = error
-      this.showError = true
-    } finally {
-      this.loading = false
-    }
-  }
-}
-</script>
