+++ pages/debug/payouts/fetch.vue
@@ -4,11 +4,32 @@
       <v-col cols="12" md="4">
         <v-form>
           <header>Optional filter params:</header>
-          <v-text-field v-model="formData.source" label="Source Wallet ID" />
           <v-text-field
-            v-model="formData.destination"
-            label="Destination Account"
+            v-model="formData.sourceWalletId"
+            label="Source Wallet ID"
           />
+          <v-text-field v-model="formData.destination" label="Destination" />
+          <v-select
+            v-model="formData.destinationType"
+            :items="destinationType"
+            label="Destination Type"
+          />
+          <v-select
+            v-model="formData.status"
+            :items="payoutStatuses"
+            label="Status"
+          />
+          <div v-if="isCryptoPayout()">
+            <v-text-field
+              v-model="formData.sourceCurrency"
+              label="Source Currency"
+            />
+            <v-text-field
+              v-model="formData.destinationCurrency"
+              label="Destination Currency"
+            />
+            <v-text-field v-model="formData.chain" label="Chain" />
+          </div>
           <v-text-field v-model="formData.from" label="From" />
           <v-text-field v-model="formData.to" label="To" />
           <v-text-field v-model="formData.pageSize" label="PageSize" />
@@ -43,8 +64,8 @@
 <script lang="ts">
 import { Component, Vue } from 'nuxt-property-decorator'
 import { mapGetters } from 'vuex'
-import RequestInfo from '@/components/RequestInfo.vue'
-import ErrorSheet from '@/components/ErrorSheet.vue'
+import RequestInfo from '~/components/RequestInfo.vue'
+import ErrorSheet from '~/components/ErrorSheet.vue'
 @Component({
   components: {
     RequestInfo,
@@ -61,8 +82,13 @@ import ErrorSheet from '@/components/ErrorSheet.vue'
 export default class FetchPayoutsClass extends Vue {
   // data
   formData = {
-    source: '',
+    sourceWalletId: '',
     destination: '',
+    destinationType: '',
+    sourceCurrency: '',
+    destinationCurrency: '',
+    chain: '',
+    status: '',
     from: '',
     to: '',
     pageSize: '',
@@ -76,10 +102,18 @@ export default class FetchPayoutsClass extends Vue {
     required: (v: string) => !!v || 'Field is required',
   }
 
+  destinationType = ['', 'wire', 'ach', 'sepa', 'address_book']
+  fiatDestinationTypes = new Set(['wire', 'ach', 'sepa'])
+  blockchainDestinationTypes = new Set(['address_book'])
+  payoutStatuses = ['', 'pending', 'complete', 'failed']
   error = {}
   loading = false
   showError = false
   // methods
+  isCryptoPayout() {
+    return this.blockchainDestinationTypes.has(this.formData.destinationType)
+  }
+
   onErrorSheetClosed() {
     this.error = {}
     this.showError = false
@@ -87,10 +121,16 @@ export default class FetchPayoutsClass extends Vue {
 
   async makeApiCall() {
     this.loading = true
+    const isCryptoPayout = this.isCryptoPayout()
     try {
       await this.$payoutsApi.getPayouts(
-        this.formData.source,
+        this.formData.sourceWalletId,
         this.formData.destination,
+        this.formData.destinationType,
+        this.formData.status,
+        isCryptoPayout ? this.formData.sourceCurrency : '',
+        isCryptoPayout ? this.formData.destinationCurrency : '',
+        isCryptoPayout ? this.formData.chain : '',
         this.formData.from,
         this.formData.to,
         this.formData.pageBefore,
